+++ public/data/abhishekkumar62000.json
@@ -4,7 +4,19 @@
   "bio": "A passion of become a Data Scientists & Artificial Intelligence, Machine Learning learn about technology, Computer Science and programming language. I am always professional in my field.",
   "avatar": "https://github.com/abhishekkumar62000.png",
   "portfolio": "https://github.com/abhishekkumar62000",
-  "skills": ["Data Scientists", "Data Analytics", "Python", "Statistics", "Power BI", "Tableau", "SQL", "Machine Learning", "Deep learning", "Generative AI", "Big Data"],
+  "skills": [
+    "Data Scientists",
+    "Data Analytics",
+    "Python",
+    "Statistics",
+    "Power BI",
+    "Tableau",
+    "SQL",
+    "Machine Learning",
+    "Deep learning",
+    "Generative AI",
+    "Big Data"
+  ],
   "social": {
     "GitHub": "https://github.com/abhishekkumar62000",
     "Twitter": "https://twitter.com/",
