+++ public/data/amanpoddar-dev12.json
@@ -4,7 +4,23 @@
   "bio": "A talented BCA student and front-end developer skilled in technologies like React, HTML, CSS, and JavaScript, with experience using frameworks like Tailwind and Bootstrap. I have successfully created numerous projects, showcasing my ability to design and build responsive, user-friendly web applications.",
   "avatar": "https://github.com/amanpoddar-dev12.png",
   "portfolio": "https://github.com/amanpoddar-dev12",
-  "skills": ["React", "JavaScript", "Tailwind", "Python", "Java", "C", "C++", "HTML", "CSS", "DBMS", "WebDevelopment", "MernStack", "Bootstrap", "PHP", "mySQL"],
+  "skills": [
+    "React",
+    "JavaScript",
+    "Tailwind",
+    "Python",
+    "Java",
+    "C",
+    "C++",
+    "HTML",
+    "CSS",
+    "DBMS",
+    "WebDevelopment",
+    "MernStack",
+    "Bootstrap",
+    "PHP",
+    "mySQL"
+  ],
   "social": {
     "GitHub": "https://github.com/amanpoddar-dev12",
     "Twitter": "https://twitter.com/AmanPoddarr",
