+++ public/data/codeaashu.json
@@ -4,7 +4,35 @@
   "bio": "A proficient Web Developer & Designer specializing in crafting captivating Websites & Apps. My focus on user-centric design, creative development, and effective collaboration, i can also elevate the brand presence with my design expertise. Till now i build 50+ WEB APP. Most importantly, I'm a reliable designer you can rely on for all your design needs.",
   "avatar": "https://github.com/codeaashu.png",
   "portfolio": "https://ashutosh-kumar.vercel.app",
-  "skills": ["React", "Tailwind", "JavaScript", "TypeScript", "NodeJS", "ThreeJS", "NextJS", "AnimeJS", "ExpressJS", "Redux", "GSAP", "MaterialUI", "SASS", "WebDevelopment", "MernStack", "WebDesign", "Figma", "Webflow", "Wix", "framer", "Bootstrap", "PHP", "mySQL", "MongoDB", "SEO", "Shell", "prompt Engineering"],
+  "skills": [
+    "React",
+    "Tailwind",
+    "JavaScript",
+    "TypeScript",
+    "NodeJS",
+    "ThreeJS",
+    "NextJS",
+    "AnimeJS",
+    "ExpressJS",
+    "Redux",
+    "GSAP",
+    "MaterialUI",
+    "SASS",
+    "WebDevelopment",
+    "MernStack",
+    "WebDesign",
+    "Figma",
+    "Webflow",
+    "Wix",
+    "framer",
+    "Bootstrap",
+    "PHP",
+    "mySQL",
+    "MongoDB",
+    "SEO",
+    "Shell",
+    "prompt Engineering"
+  ],
   "social": {
     "GitHub": "https://github.com/codeaashu",
     "Twitter": "https://twitter.com/warrior_aashuu",
