+++ public/data/deruxhalter3js.json
@@ -19,7 +19,8 @@
     "Git",
     "Tailwind CSS",
     "REST APIs",
-    "GraphQL"],
+    "GraphQL"
+  ],
   "social": {
     "GitHub": "https://github.com/deruxhalter3js",
     "Twitter": "https://twitter.com/DavidMwas2_0",
