+++ public/data/shubhcoding01.json
@@ -4,7 +4,30 @@
   "bio": "I'm Web-Developer, I'm currently working as Freelancer, Open to work as a Software Engineer, I'm currently learning MERN, Funfact: I also do Photo/Video Editing ",
   "avatar": "https://github.com/shubhcoding01.png",
   "portfolio": "https://github.com/shubhcoding01",
-  "skills": ["Java", "JavaScript", "HTML", "CSS", "C++", "DSA", "MY SQL", "PYTHON", "C", "React", "Cyber Security", "WebFlow", "Canva", "Figma", "NodeJS", "WebDevelopment", "FullStack", "MernStack", "GSAP", "Git", "GitHub", "KALI LINUX"],
+  "skills": [
+    "Java",
+    "JavaScript",
+    "HTML",
+    "CSS",
+    "C++",
+    "DSA",
+    "MY SQL",
+    "PYTHON",
+    "C",
+    "React",
+    "Cyber Security",
+    "WebFlow",
+    "Canva",
+    "Figma",
+    "NodeJS",
+    "WebDevelopment",
+    "FullStack",
+    "MernStack",
+    "GSAP",
+    "Git",
+    "GitHub",
+    "KALI LINUX"
+  ],
   "social": {
     "GitHub": "https://github.com/shubhcoding01",
     "Twitter": "https://x.com/shubh__coding",
