+++ src/components/Sidebar/AboutUs.jsx
@@ -1,0 +1,3 @@
+export default function AboutUS() {
+  return <h1>ABout us</h1>;
+}
