+++ src/components/Sidebar/Sidebar.jsx
@@ -1,10 +1,13 @@
 import React, { useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCode, faMoon, faSun } from '@fortawesome/free-solid-svg-icons';
+import { FaLinkedin } from 'react-icons/fa';
+import { FcAbout } from 'react-icons/fc';
 
+import { useNavigate } from 'react-router-dom';
 function Sidebar() {
   const [theme, setTheme] = useState('dark');
-
+  const [showAbout, setShowAbout] = useState(false);
   function toggleTheme() {
     const htmlElement = document.documentElement;
     const isDarkModeEnabled = htmlElement.classList.contains('dark');
@@ -17,7 +20,12 @@ function Sidebar() {
       setTheme('dark');
     }
   }
-
+  const navigate = useNavigate();
+  function handleAbout(e) {
+    e.preventDefault();
+    setShowAbout(() => (showAbout ? false : true));
+    navigate('/about');
+  }
   return (
     <div className="my-7 w-full border-r-2 border-borderSecondary px-7 font-spaceMono dark:border-borderColor md:h-[90vh] md:w-[23%] md:px-2 lg:px-7">
       <div className="mb-2 flex h-12 items-center gap-2.5">
@@ -45,20 +53,28 @@ function Sidebar() {
         </div>
       </div>
       <div className="text-secondaryColor dark:text-white">Discover and Connect with Skilled Developers.</div>
-      <div className="pt-5">
+      <div className="my-5 flex flex-row items-center justify-center space-x-3 pt-5 md:flex-col md:space-x-0 md:space-y-5">
         <a href="https://github.com/codeaashu/DevDisplay#how-to-add-your-profile-" target="_blank" rel="noreferrer">
-          <button className="inline-block cursor-pointer rounded-lg border-2 border-textSecondary bg-textSecondary px-[15px] py-1.5 text-center font-poppoins text-sm transition-all duration-500 hover:bg-transparent hover:text-textSecondary dark:text-white mr-4">
+          <button className="inline-block h-[50px] w-full cursor-pointer rounded-lg border-2 border-textSecondary bg-textSecondary px-6 py-2 text-center font-poppoins text-sm transition-all duration-500 hover:bg-transparent hover:text-textSecondary dark:text-white md:w-[200px]">
             Add your profile
           </button>
         </a>
+
         <a href="https://www.linkedin.com/company/devdisplay/" target="_blank" rel="noreferrer">
-          <button className="inline-block cursor-pointer rounded-lg border-2 border-textSecondary bg-textSecondary px-[15px] py-1.5 text-center font-poppoins text-sm transition-all duration-500 hover:bg-transparent hover:text-textSecondary dark:text-white">
-            Connect us
+          <button className="mt-0 inline-flex h-[50px] w-full cursor-pointer items-center justify-center space-x-3 rounded-lg border-2 border-textSecondary bg-textSecondary px-6 py-2 font-poppoins text-sm transition-all duration-500 hover:bg-transparent hover:text-textSecondary dark:text-white md:w-[200px]">
+            <span>Connect us</span>
+            <FaLinkedin className="text-[1.2rem] text-blue-600 duration-300 hover:scale-125" />
           </button>
         </a>
       </div>
+      <button
+        className="border-secondaryColor-blue mt-3 inline-flex h-[50px] w-full items-center justify-center space-x-3 rounded-lg border-2 border-blue-400 bg-secondaryColor px-6 py-2 font-poppoins text-sm text-white transition-all duration-500 dark:border-textSecondary dark:text-white md:mx-12 md:mt-0 md:flex md:w-[200px] md:space-x-3"
+        onClick={handleAbout}
+      >
+        <span className="md:flex md:items-center">About Us</span>
+        <FcAbout className="text-[1.2rem] duration-300 hover:scale-125" />
+      </button>
     </div>
   );
 }
-
 export default Sidebar;
